color mode for random spawn renderer

+++ src/components/Settings/Range.tsx
@@ -21,7 +21,7 @@ export const Range: React.FC<RangeProps> = ({
             min={min}
             step={step}
             value={value}
-            onChange={(e) => onChange(parseInt(e.target.value))}
+            onChange={(e) => onChange(parseFloat(e.target.value))}
         />
     );
 };
